sdm: Operate FINGERPRINT_MASK for oplus optical UDFPS
Even though oplus calls the connector property hbm_enable, it's
function is not just limited to operating HBM. It is also responsible
for other aspects of UDFPS operations like that of setFpPress so that
we don't have to handle any of these through IOCTL.

The purpose of this DRM connector property is for the userspace to let
the kernel display and fingerprint drivers know which layers are visible.
The same can be achieved by asserting with the layer name instead of setting
a magic constant for the layer ZOrder, thereby eliminating the need for
SurfaceFlinger hacks that we previously needed with prebuilt HWC.

| Bit index | Decimal | Layer       | Z-order      |
| --------- | ------- | ----------- | ------------ |
| 0         | 1       | **dim**     | `0x41000005` |
| 1         | 2       | **pressed** | `0x41000033` |
| 2         | 4       | **icon**    | `0x41000032` |
| 3         | 8       | **aod**     | `0x41000011` |

+++ sde-drm/drm_connector.cpp
@@ -981,7 +981,7 @@ void DRMConnector::Perform(DRMOps code, drmModeAtomicReq *req, va_list args) {
                  __FUNCTION__, obj_id);
         return;
       }
-      uint32_t fingerprint_mask = va_arg(args, uint32_t);
+      uint8_t fingerprint_mask = va_arg(args, uint32_t);
       uint32_t prop_id = prop_mgr_.GetPropertyId(DRMProperty::FINGERPRINT_MASK);
       int ret = drmModeAtomicAddProperty(req, obj_id, prop_id, fingerprint_mask);
       if (ret < 0) {
